Move conda functions in a conda utils python file
