+++ mlflow/pyfunc/backend.py
@@ -8,8 +8,7 @@ from mlflow.models.docker_utils import _build_image, DISABLE_ENV_CREATION
 from mlflow.pyfunc import ENV
 from mlflow.pyfunc import scoring_server
 
-from mlflow.projects.utils import get_or_create_conda_env, get_conda_bin_executable, \
-    get_conda_command
+from mlflow.utils.conda import get_or_create_conda_env, get_conda_bin_executable, get_conda_command
 from mlflow.tracking.artifact_utils import _download_artifact_from_uri
 from mlflow.utils.file_utils import path_to_local_file_uri
 from mlflow.version import VERSION
